(Arch Linux) Add wireshark/config.h from ubuntu deb package and change version number.
To build the dissector plugin under Arch Linux, the config.h file is
needed. I copied it from the ubuntu package, put in the current version
number and it worked! It is also necessary to update the install path in
CMakeLists.txt so that wireshark finds and loads the plugin.